client side error handling DONE

+++ src/Resources/Pages/Admin/Media/Media.js
@@ -20,21 +20,29 @@ module.exports = {
 	<div id="media-display-cont">
 
 	${async () => {
-    return String(
-      (await Index.pool.query(`SELECT * FROM media`)).rows.map((t) => {
-        return `
-	
-		<div data-media-id="${t.id}" data-media-full-url="${t.full_url}" data-media-alt-text="${t.alt_text}"  class="media-element">
 
-		<img src="${t.full_url}" alt="${t.alt_text}" />
+		var records = await Index.pool.query(`SELECT * FROM media`)
 
-		<a target="_blank" href="${t.full_url}">${t.full_url}</a>
+		if(records.rowCount != 0){
+			return String(
+				records.rows.map((t) => {
+				  return `
+			  
+				  <div data-media-id="${t.id}" data-media-full-url="${t.full_url}" data-media-alt-text="${t.alt_text}"  class="media-element">
+		  
+				  <img src="${t.full_url}" alt="${t.alt_text}" />
+		  
+				  <a target="_blank" href="${t.full_url}">${t.full_url}</a>
+		  
+				  <button class="media-delete">delete</button>
+			  </div>
+				  `
+				}),
+			  ).replaceAll(",", "\n")
+		}else{
+			return `<h2>No Media Found</h2>`
+		}
 
-		<button class="media-delete">delete</button>
-	</div>
-		`;
-      }),
-    ).replaceAll(",", "\n");
   }}
 
 
@@ -77,40 +85,61 @@ gap:.5rem;
 	</style>
 
 		<script>
-	document.querySelectorAll(".nav-menu__item").forEach( (node, index) => { node.classList.remove("is-active")})
-	document.querySelector(".navbar-media").classList.add("is-active")
+
+		document.querySelectorAll(".nav-menu__item").forEach( (node, index) => { node.classList.remove("is-active")})
+document.querySelector(".navbar-media").classList.add("is-active")
 
 document.querySelectorAll(".media-delete").forEach(t => {t.addEventListener("click", async (e)=> {
 
 
-	const formData = new FormData();
- 
-	
-
-	formData.append("id", e.target.parentNode.getAttribute("data-media-id"))
- 
-		document.querySelector(".loading-block").classList.add("active")
-
- 	const res = await fetch(".", {
-
-	method:"DELETE",
-	body: formData,
-
-	
-	
-	})
-
-			document.querySelector(".loading-block").classList.remove("active")
-
-			if(res.ok){
-			
-			window.location.href = "."
-			
-			}
+try{
+if (window.navigator.onLine) {
+const formData = new FormData();
 
 
-	})})
-	
+formData.append("id", e.target.parentNode.getAttribute("data-media-id"))
+
+	document.querySelector(".loading-block").classList.add("active")
+
+ const res = await fetch(".", {
+
+method:"DELETE",
+body: formData,
+
+
+
+})
+
+		document.querySelector(".loading-block").classList.remove("active")
+
+		if(res.ok){
+		
+		window.location.href = "."
+		
+		}else{
+		
+			document.querySelector("#qMQEbc-container").classList.add("active")
+document.querySelector("#qMQEbc-message").innerHTML = "Unknown Error"}
+
+		}
+
+
+
+else{
+document.querySelector("#qMQEbc-container").classList.add("active")
+document.querySelector("#qMQEbc-message").innerHTML = "No Internet Connection"}
+
+}
+
+catch(e){
+
+document.querySelector("#qMQEbc-container").classList.add("active")
+document.querySelector("#qMQEbc-message").innerHTML = "Unknown Error"
+}
+
+})})
+
+
 	</script>
 	
 
@@ -119,3 +148,8 @@ document.querySelectorAll(".media-delete").forEach(t => {t.addEventListener("cli
   `,
     }),
 };
+ 
+	
+
+
+
